Create spoke_cli.sh
This is straightforward globus cli syntax to run a set of transfers from a script.  It requires that the endpoints be activated in your Globus Online web interface, or it will prompt vie the globus cli to activate them.   Doing small test transfers seems to be the best way to know for sure that Globus is functional with an endpoint.